refactor: 참조관계를 고려하여 반환 받은 response를 toInstance에서 static from으로 변경
