+++ src/main/java/com/programmers/bucketback/domains/review/api/dto/response/ReviewGetByCursorResponse.java
@@ -2,6 +2,7 @@ package com.programmers.bucketback.domains.review.api.dto.response;
 
 import java.util.List;
 
+import com.programmers.bucketback.domains.review.application.dto.ReviewGetByCursorServiceResponse;
 import com.programmers.bucketback.domains.review.application.vo.ReviewCursorSummary;
 
 public record ReviewGetByCursorResponse(
@@ -9,4 +10,11 @@ public record ReviewGetByCursorResponse(
 	String nextCursorId,
 	List<ReviewCursorSummary> reviews
 ) {
+	public static ReviewGetByCursorResponse from(final ReviewGetByCursorServiceResponse response) {
+		return new ReviewGetByCursorResponse(
+			response.reviewCount(),
+			response.nextCursorId(),
+			response.reviews()
+		);
+	}
 }
