+++ src/main/java/com/programmers/bucketback/domains/review/application/dto/ReviewGetByCursorServiceResponse.java
@@ -2,7 +2,6 @@ package com.programmers.bucketback.domains.review.application.dto;
 
 import java.util.List;
 
-import com.programmers.bucketback.domains.review.api.dto.response.ReviewGetByCursorResponse;
 import com.programmers.bucketback.domains.review.application.vo.ReviewCursorSummary;
 
 public record ReviewGetByCursorServiceResponse(
@@ -10,11 +9,4 @@ public record ReviewGetByCursorServiceResponse(
 	String nextCursorId,
 	List<ReviewCursorSummary> reviews
 ) {
-	public ReviewGetByCursorResponse toReviewGetByCursorResponse() {
-		return new ReviewGetByCursorResponse(
-			reviewCount,
-			nextCursorId,
-			reviews
-		);
-	}
 }
